feat: add more validation on deleteFileFromCache

+++ features/content/play_common/src/main/java/com/tokopedia/play_common/util/CacheUtil.kt
@@ -8,7 +8,7 @@ import com.tokopedia.utils.file.FileUtil
 object CacheUtil {
 
     fun deleteFileFromCache(filePath: String) {
-        if (filePath.contains(FileUtil.getTokopediaInternalDirectory(null).path)) {
+        if (filePath.isNotBlank() && filePath.contains(FileUtil.getTokopediaInternalDirectory(null).path)) {
             FileUtil.deleteFile(filePath)
         }
     }
